work/070-django-rest/requirements.txt
django<3
django-registration
djangorestframework
requests
django-rest-swagger
